+++ src/envoy-node-boilerplate.ts
@@ -1,13 +1,6 @@
-export * from "./envoy-context";
-export * from "./envoy-fetch";
-export * from "./envoy-grpc-request-params";
-export * from "./envoy-http-client";
-export * from "./envoy-http-request-params";
-export * from "./envoy-request-params";
-
 export { default as EnvoyContext } from "./envoy-context";
 export { default as envoyFetch } from "./envoy-fetch";
-export { default as EnvoyGrpcRequestParams } from "./envoy-grpc-request-params";
+export { default as EnvoyGrpcRequestParams, GrpcRetryOn } from "./envoy-grpc-request-params";
 export { default as EnvoyHttpClient } from "./envoy-http-client";
-export { default as EnvoyHttpRequestParams } from "./envoy-http-request-params";
+export { default as EnvoyHttpRequestParams, HttpRetryOn } from "./envoy-http-request-params";
 export { default as envoyProtoDecorator } from "./envoy-proto-decorator";
