+++ test/grpc.test.ts
@@ -3,8 +3,7 @@ import grpc, { ServerUnaryCall, sendUnaryData, ServiceError } from "grpc";
 import GrpcTestServer, { Ping, PingEnvoyClient } from "./lib/grpc-test-server";
 import { sleep } from "./lib/utils";
 import { RequestFunc, EnvoyClient } from "../src/types";
-import EnvoyContext from "../src/envoy-context";
-import { GrpcRetryOn } from "../src/envoy-node-boilerplate";
+import { GrpcRetryOn, EnvoyContext } from "../src/envoy-node-boilerplate";
 
 interface PingEnvoyClient extends EnvoyClient {
   inner: RequestFunc;
